SampleTunerTIS setup read SectionEvent data
In the SetupActivity, a sectionFilter is created in order
to receive SectionEvents which can provide us channel
and network data. In this CL, SetupActivity is modified
to remain open until it receives data from a SectionEvent.
Currently it only expects a single test value which is sets
as its channel number.

Bug: 237323181
Test: Run manually on Cuttlefish using new test.es
Test: containing section data. Should see channel number
Test: matching input file, with no other functionality changes.
Change-Id: Icca8c0dc605fc83df00ad99e65f687fd477a646d