SampleTunerTIS Add SectionParser for VCT
The SetupActivity for the SampleTunerTIS is now receiving
SectionEvents. This CL adds a SectionParser in order to
process these events, as well as the method and data
structure which will be needed for the specific VCT event
which we currently expect.

Bug: 237323181
Test: Run manually on Cuttlefish. Should see a received
event, and see the default channel number and name
being received from the parsing method.

Change-Id: I8dbd0b0535d9caa3d685aa9b2865d954315d4cd9