SampleTunerTIS add descriptor parsing
When we are parsing the TVCT table in our SectionParser,
we require some data which is located within descriptors.
This CL adds a framework for parsing descriptors, which
we will utilize in future CLs to extract the
ExtendedChannelName for our SetupActivity.

Bug: 237323181
Test: Run manually with cuttlefish. Should see the channel
name "Sample" if an ExtendedChannelNameDescriptor exists,
and the shortName located in the TVCT otherwise.

Change-Id: Ife8418642169d0de7cca85cf679a30a452c69716