SampleTunerTIS Add sleeping between media frames
Currently, the TIS queues frames into the MediaCodec as
quickly as it can, ignoring their Presentation Time Stamps.
This CL instead has it properly wait in between each
frame until the correct time, producing a smooth output
of the correct length.

Bug: 238807604
Test: Manually using Cuttlefish and video input. Compare
the elapsed SystemTime vs. the known length of input.

Change-Id: I2cef080a46b5e5d547c2cbe1e3d69811ee13282a