SampleTunerTIS Fix bug with mDataQueue overfilling
Currently, when SAVE_DATA is enabled, all saved data is
added to the mDataQueue on every loop of our thread. This
results in the mDataQueue's size rapidly growing as the
TIS runs. This CL changes the behavior to instead fill the
mDataQueue only if it is currently empty, and adds logging
for when it occurs.

Bug: 238929236
Test: Manually with Cuttlefish.
Change-Id: Ie928e58352ecc093ca8427e3f9c1b56263000e64