SampleTunerTIS Create SectionFilterCallback in TIS
Currently, the TIS uses a default logging callback to
handle incoming SectionEvents. This CL replaces the default
callback with a custom one that calls a new function to
handle the events. Implementation of the handleSection
function is left to future CLs.

Bug: 238890289
Test: Manually with Cuttlefish and input including sections.
AV input handling is not changed while section size is logged.

Change-Id: I8669e94b88c3f34f0181b46974b6b92909d1f950