SampleTunerTIS Insert program from EitEventInfo
This CL takes the EitEventInfo which [will be] parsed
inside of parseEitSection() and inserts it into the program
information of the running TV Application. Before
insertion, it removes all programs which would conflict
with its running time.

Bug: 238890289
Test: Manually with Cuttlefish. Should see created program
with sample title and length. Repeatedly opening and closing
the channel will see the program replaced to match current
starting time.

Change-Id: Ibec1e259e7063a5edb62f921c63f3d5cc0315c02