SampleTunerTIS Seperate parseMultiString
Previously, we implemented a
parseExtendedChannelNameDescriptor method. However, this
descriptor is defined as simply a tag and length, followed
by a MultipleStringStructure. This CL moves the parsing of
this string structure into its own seperate method so that
it can be used in the parsing of EIT sections.

Bug: 238890289
Test: Manually with Cuttlefish. Using bad string structures
in existing ExtendedChannelNameDescriptors should result in
new error logs, with no other functionality changes.

Change-Id: I8367b8467381efce8f430de4abf3cdbe4dfea933