SampleTunerTIS fix program initialization in setup
During the SetupActivity of the Sample TIS, the Channel's
information is set with a mix of data acquired from the VCT
and hardcoded values. One of these hardcoded values was an
initial program. However, the ChannelUtils.updateChannels
method called actually entirely ignored this program.
Therefore it is replaced with a call to ProgramUtils
which autogenerates programs to fill an entire day of
scheduling.

Bug: 239457222
Test: Run manually with Cuttlefish. Should see test programs
filling up 24 hours of scheduling. Running with TIS EIT
changes will see overlapping programs deleted when EIT is read.

Change-Id: Ieb67545fc92c1eda27235783aea1bd54192d7186