SampleTunerTIS Store MediaEventData to allow recycling MediaEvents
The SampleTunerTIS receives MediaEvents from its filter
which are originally created through the tuner JNI code.
In our current code, the recycle() method is never called
on the LinearBlocks generated by the native code, which
causes a memory leak that leads to crashes. This CL fixes
this problem by making a new MediaEventData java class
which we use to store the data which is normally held
in MediaEvents. This allows us to recycle MediaEvents
as soon as we receieve them, allowing our JNI code to
deallocate the memory they would otherwise take up.

Bug: 239732425
Test: Manually with Cuttlefish. Memory usage no longer
stays elevated after TIS is closed, and crashes are fixed.

Change-Id: Ic4689078493611c150d5b4eb10b4ed7cc1123bea