SampleTIAS Create surface display
In this CL, the sample TIAS responds to onSetSurface and
onSurfaceChanged by creating a virtualDisplay which it
fills with the new mViewContainer. This container is
currently only a LinearLayout which has a red background
for testing purposes, however in future CLs it will
be filled with proper GUI views.

Bug: 240203641
Test: Manually with Cuttlefish. After TIS calls the TIAS,
should see screen entirely filled with the red background.

Change-Id: I051588da3b953b3fa52aa34fda70bbcc86f5efc4