SampleTIAS Add tuning on button press
Currently, our SampleTIAS requests data and creates an
overlay for itself upon initialization. This CL adds its
first piece of user interactivity by allowing the user
to press the Red Interactive Button in order to tune to
the next channel. A field is added to the GUI to inform
the user of this option.

Bug: 240578538
Image: https://screenshot.googleplex.com/3TZxVRSTfqVspar.png
Test: Manually with Cuttlefish, LiveTv, and SampleTunerTIS
Change-Id: Ibf34a494bbf8b610036aff5d4da04260da632af1