SampleTIAS Add selected track info to overlay
In the SampleTIAS, an overlay is used to display the current
value of important variables. While previously we only had
initial default values for the selected tracks, this CL
adds their real values to the overlay. An initial default
value is still needed, however, since we cannot request the
data for which tracks are selected and must instead wait
for it to be sent to us.

Bug: 240203641
Image: https://screenshot.googleplex.com/4BYTLpuZsUkdDWu.png
Test: Manually with Cuttlefish, LiveTv, and SampleTunerTIS
Change-Id: I85a0ab80f56a0979b73500c035cd1f0e5dcd8235