LiveTv TIAS Integration playbackCommandRequest
The TvInteractiveAppView.TvInteractiveAppCallback includes
a playbackCommandRequest method which allows the TIAS to
request changes to be made on the application running it.
This CL implements all of these commands except track
selection, which will be implemented in a future CL
alongside changes to the audio system needed to
support it.

Bug: 241110408
Test: Manually with Cuttlefish and SampleTIAS
Change-Id: I6e9f9065dd28dad4451150daa5fe5deb5b3b38d1