SampleTunerTIS Correct notifyVideoAvailable use
The notifyVideoAvailable() function of the TIS is used to
signal that the content rendered onto its surface is ready
to be viewed, and must be called at least once per call to
onTune(). After it is called, the manager will assume that
the video is still available until notiftVideoUnavailable()
is called.

Our current implementation of the SampleTunerTIS calls
notifyVideoAvailable() for each frame of video played.
This CL changes the behavior to only call the method once
per call to onTune, as well as notifying the video as
unavailable while we do our setting up/tuning.

Bug: 242907861
Test: Manually with Cuttlefish and LiveTv.
Change-Id: I9d8d623a9b8e8fe1c3199672cfcd57599a7f572b