SampleTunerTIS Choose between Dvr and FE input
Currently, the SampleTunerTIS uses a combination of
frontend tuning and dvr reading to receive its input. This
combination does not match real use cases and causes an
incorrect state in the Tuner HAL, as described in
attached bug. This CL changes the Tuner to instead use
only the DVR or Frontend input depending on a static
flag given at compile time.

Note that with the current state of the default Tuner HAL,
only DVR playback mode will result in input being fed to
TIS. The inclusion of Frontend Input still serves as a
valuable sample of a real use case, and can be implemented
through changes to our default Tuner HAL.

Bug: 239731089
Test: Manually with Cuttlefish and LiveTv
Change-Id: I69ce32a9732dce664cd79381a4634e905ab5bd63