Removed getAppManifest method
As config.resourceDir() and config.assetDir() has been removed from
library. The legacy way of programmatically constructing an
AndroidManifest should be updated to Android.bp in Soong
In addition, there seems to be no usage of this method in git_main.

This is necessary as the latest update from Robolectric library has removed this deprecated method. The update ag/35465139 is currently blocked by this update.

Test: atest TvRoboTests
Bug: 443072871

Flag: EXEMPT TEST_ONLY

Change-Id: Idc185f426deefc43c1f1c69a361c3838f0523d86